requirements.txt
kivy==2.2.1
opencv-python==4.8.1.78
face-recognition==1.3.5
face-detect-recognition==0.2.0
numpy==1.24.3
pillow==10.0.0
buildozer==1.4.6
cython==0.29.36
